move tls-related code to `pkg/controller/ps/tls.go`
