+++ pkg/controller/ps/controller.go
@@ -27,8 +27,6 @@ import (
 	"strings"
 	"time"
 
-	cm "github.com/cert-manager/cert-manager/pkg/apis/certmanager/v1"
-	cmmeta "github.com/cert-manager/cert-manager/pkg/apis/meta/v1"
 	"github.com/pkg/errors"
 	"golang.org/x/sync/errgroup"
 	appsv1 "k8s.io/api/apps/v1"
@@ -609,28 +607,6 @@ func (r *PerconaServerMySQLReconciler) reconcileUsers(ctx context.Context, cr *a
 	}
 
 	log.Info("Updated internal secret", "secretName", cr.InternalSecretName())
-
-	return nil
-}
-
-func (r *PerconaServerMySQLReconciler) ensureTLSSecret(
-	ctx context.Context,
-	cr *apiv1alpha1.PerconaServerMySQL,
-) error {
-	err := r.ensureSSLByCertManager(ctx, cr)
-	if err != nil {
-		if cr.Spec.TLS != nil && cr.Spec.TLS.IssuerConf != nil {
-			return errors.Wrap(err, "create ssl with cert manager")
-		}
-		secret, err := secret.GenerateCertsSecret(ctx, cr)
-		if err != nil {
-			return errors.Wrap(err, "create SSL manually")
-		}
-
-		if err := k8s.EnsureObjectWithHash(ctx, r.Client, cr, secret, r.Scheme); err != nil {
-			return errors.Wrap(err, "create secret")
-		}
-	}
 
 	return nil
 }
@@ -1717,162 +1693,6 @@ func (r *PerconaServerMySQLReconciler) restartGroupReplication(ctx context.Conte
 		return errors.Wrapf(err, "start group replication on %s", primary)
 	}
 	log.V(1).Info("Started group replication", "hostname", primary)
-
-	return nil
-}
-
-func (r *PerconaServerMySQLReconciler) ensureSSLByCertManager(ctx context.Context, cr *apiv1alpha1.PerconaServerMySQL) error {
-	issuerName := cr.Name + "-pso-issuer"
-	caIssuerName := cr.Name + "-pso-ca-issuer"
-	issuerKind := "Issuer"
-	issuerGroup := ""
-	if cr.Spec.TLS != nil && cr.Spec.TLS.IssuerConf != nil {
-		issuerKind = cr.Spec.TLS.IssuerConf.Kind
-		issuerName = cr.Spec.TLS.IssuerConf.Name
-		issuerGroup = cr.Spec.TLS.IssuerConf.Group
-	} else {
-		issuerConf := cm.IssuerConfig{
-			SelfSigned: &cm.SelfSignedIssuer{},
-		}
-		if cr.Spec.TLS != nil && cr.Spec.TLS.IssuerConf != nil {
-			issuerConf = cm.IssuerConfig{
-				CA: &cm.CAIssuer{SecretName: cr.Spec.TLS.IssuerConf.Name},
-			}
-		}
-		if err := r.ensureIssuer(ctx, cr, caIssuerName, issuerConf); err != nil {
-			return err
-		}
-
-		caCert := &cm.Certificate{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      cr.Name + "-ca-cert",
-				Namespace: cr.Namespace,
-			},
-			Spec: cm.CertificateSpec{
-				SecretName: cr.Name + "-ca-cert",
-				CommonName: cr.Name + "-ca",
-				IsCA:       true,
-				IssuerRef: cmmeta.ObjectReference{
-					Name:  caIssuerName,
-					Kind:  issuerKind,
-					Group: issuerGroup,
-				},
-				Duration:    &metav1.Duration{Duration: time.Hour * 24 * 365},
-				RenewBefore: &metav1.Duration{Duration: 730 * time.Hour},
-			},
-		}
-		if err := k8s.EnsureObjectWithHash(ctx, r.Client, cr, caCert, r.Scheme); err != nil {
-			return errors.Wrap(err, "ensure CA certificate")
-		}
-
-		if err := r.waitForCerts(ctx, cr.Namespace, cr.Name+"-ca-cert"); err != nil {
-			return err
-		}
-
-		issuerConf = cm.IssuerConfig{
-			CA: &cm.CAIssuer{SecretName: caCert.Spec.SecretName},
-		}
-
-		if err := r.ensureIssuer(ctx, cr, issuerName, issuerConf); err != nil {
-			return err
-		}
-	}
-	hosts := []string{
-		fmt.Sprintf("*.%s-mysql", cr.Name),
-		fmt.Sprintf("*.%s-mysql.%s", cr.Name, cr.Namespace),
-		fmt.Sprintf("*.%s-mysql.%s.svc.cluster.local", cr.Name, cr.Namespace),
-		fmt.Sprintf("*.%s-orchestrator", cr.Name),
-		fmt.Sprintf("*.%s-orchestrator.%s", cr.Name, cr.Namespace),
-		fmt.Sprintf("*.%s-orchestrator.%s.svc.cluster.local", cr.Name, cr.Namespace),
-		fmt.Sprintf("*.%s-router", cr.Name),
-		fmt.Sprintf("*.%s-router.%s", cr.Name, cr.Namespace),
-		fmt.Sprintf("*.%s-router.%s.svc.cluster.local", cr.Name, cr.Namespace),
-	}
-	kubeCert := &cm.Certificate{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      cr.Name + "-ssl",
-			Namespace: cr.Namespace,
-		},
-		Spec: cm.CertificateSpec{
-			SecretName: cr.Spec.SSLSecretName,
-			DNSNames:   hosts,
-			IsCA:       false,
-			IssuerRef: cmmeta.ObjectReference{
-				Name:  issuerName,
-				Kind:  issuerKind,
-				Group: issuerGroup,
-			},
-		},
-	}
-	if cr.Spec.TLS != nil {
-		kubeCert.Spec.DNSNames = append(kubeCert.Spec.DNSNames, cr.Spec.TLS.SANs...)
-	}
-
-	if err := k8s.EnsureObjectWithHash(ctx, r.Client, cr, kubeCert, r.Scheme); err != nil {
-		return errors.Wrap(err, "ensure certificate")
-	}
-
-	return r.waitForCerts(ctx, cr.Namespace, cr.Spec.SSLSecretName)
-}
-
-func (r *PerconaServerMySQLReconciler) waitForCerts(ctx context.Context, namespace string, secretName string) error {
-	ticker := time.NewTicker(3 * time.Second)
-	timeoutTimer := time.NewTimer(30 * time.Second)
-	defer timeoutTimer.Stop()
-	defer ticker.Stop()
-	secretFound := false
-	for {
-		select {
-		case <-timeoutTimer.C:
-			if !secretFound {
-				return errors.Errorf("timeout: can't get tls certificate from certmanager: %s", secretName)
-			}
-			return errors.Errorf("timeout: tls certificate from certmanager is not ready: %s", secretName)
-		case <-ticker.C:
-			secret := new(corev1.Secret)
-			err := r.Get(ctx, types.NamespacedName{
-				Name:      secretName,
-				Namespace: namespace,
-			}, secret)
-			if client.IgnoreNotFound(err) != nil {
-				return errors.Wrap(err, "failed to get secret")
-			} else if k8serrors.IsNotFound(err) {
-				continue
-			}
-			secretFound = true
-
-			cert := new(cm.Certificate)
-			err = r.Get(ctx, types.NamespacedName{
-				Name:      secretName,
-				Namespace: namespace,
-			}, cert)
-			if err != nil {
-				return errors.Wrap(err, "failed to get certificate")
-			}
-			for _, cond := range cert.Status.Conditions {
-				if cond.Type == cm.CertificateConditionReady && cond.Status == cmmeta.ConditionTrue {
-					return nil
-				}
-			}
-		}
-	}
-}
-
-func (r *PerconaServerMySQLReconciler) ensureIssuer(ctx context.Context, cr *apiv1alpha1.PerconaServerMySQL, issuerName string, IssuerConf cm.IssuerConfig,
-) error {
-	isr := &cm.Issuer{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      issuerName,
-			Namespace: cr.Namespace,
-		},
-		Spec: cm.IssuerSpec{
-			IssuerConfig: IssuerConf,
-		},
-	}
-	err := k8s.EnsureObjectWithHash(ctx, r.Client, cr, isr, r.Scheme)
-	if err != nil {
-		return errors.Wrap(err, "create issuer")
-	}
 
 	return nil
 }
